fixed reset and cleaned up some code.
