+++ src/components/Popup.js
@@ -2,7 +2,7 @@ export default class Popup {
   constructor(popupSelector) {
     this._popupElement = document.querySelector(popupSelector);
 
-    this._popupClose = this._popupElement.querySelector(".modal__close");
+    this._popupCloseBtn = this._popupElement.querySelector(".modal__close");
 
     this._handleEscClose = this._handleEscClose.bind(this);
     this._handleModalClose = this._handleModalClose.bind(this);
@@ -13,11 +13,13 @@ export default class Popup {
     this._popupElement.classList.add("modal_opened");
     document.addEventListener("keydown", this._handleEscClose);
     this._popupElement.addEventListener("click", this._handleModalClose);
-    this._popupElement.addEventListener("click", this._handleButtonClose);
   }
 
   close() {
     this._popupElement.classList.remove("modal_opened");
+
+    document.removeEventListener("keydown", this._handleEscClose);
+    this._popupElement.removeEventListener("click", this._handleModalClose);
   }
 
   _handleOpenPopup(evt) {
@@ -38,18 +40,11 @@ export default class Popup {
     }
   }
 
-  _handleButtonClose(evt) {
-    if (evt.target === this._popupClose) {
-      this.close();
-    }
+  _handleButtonClose() {
+    this.close();
   }
 
   setEventListeners() {
-    if (this._popupElement) {
-      this._popupElement.addEventListener("click", this._handleOpenPopup);
-    }
-    document.removeEventListener("keydown", this._handleEscClose);
-    this._popupElement.removeEventListener("click", this._handleModalClose);
-    this._popupElement.removeEventListener("click", this._handleButtonClose);
+    this._popupCloseBtn.addEventListener("click", this._handleButtonClose);
   }
 }
