+++ src/index.html
@@ -101,7 +101,7 @@
             <input
               class="modal__input modal__input_type_title"
               type="text"
-              name="title"
+              name="name"
               id="modal-card-title"
               placeholder="Title"
               required
@@ -115,7 +115,7 @@
               class="modal__input modal__input_type_url"
               type="url"
               id="modal-url-input"
-              name="URL"
+              name="link"
               placeholder="Image URL"
               required
             />
